added package-info.java javadoc and relocated a class
