+++ core/src/test/java/org/apache/sling/testing/mock/osgi/config/ConfigTypeContextImplTest.java
@@ -19,7 +19,6 @@
 package org.apache.sling.testing.mock.osgi.config;
 
 import org.apache.sling.testing.mock.osgi.MapUtil;
-import org.apache.sling.testing.mock.osgi.config.annotations.ConfigAnnotationUtil;
 import org.apache.sling.testing.mock.osgi.config.annotations.ConfigType;
 import org.apache.sling.testing.mock.osgi.config.annotations.TypedConfig;
 import org.apache.sling.testing.mock.osgi.config.annotations.UpdateConfig;
