+++ junit5/src/main/java/org/apache/sling/testing/mock/osgi/junit5/OsgiConfigParametersExtension.java
@@ -19,7 +19,7 @@
 package org.apache.sling.testing.mock.osgi.junit5;
 
 import org.apache.sling.testing.mock.osgi.config.ConfigTypeContext;
-import org.apache.sling.testing.mock.osgi.config.annotations.ConfigAnnotationUtil;
+import org.apache.sling.testing.mock.osgi.config.ConfigAnnotationUtil;
 import org.apache.sling.testing.mock.osgi.config.annotations.ConfigCollection;
 import org.apache.sling.testing.mock.osgi.config.annotations.ConfigType;
 import org.apache.sling.testing.mock.osgi.config.annotations.UpdateConfig;
